test al 70 %
